% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czdaptools-package.r
\docType{package}
\name{czdaptools}
\alias{czdaptools}
\alias{czdaptools-package}
\title{Tools for downloading zone data from ICANN's CZDS application.}
\description{
Can perform parallel downloads. Use \code{options(mc.cores=n)}. Also expects
the ICANN CZDAP key to be in ICANN_CZDAP_KEY (use \code{.Renvion}).
}
\note{
download & decryption of "credentials" it not supported yet
}
\author{
Bob Rudis (@hrbrmstr)
}

