% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czdaptools.r
\name{get_zone_data_urls}
\alias{get_zone_data_urls}
\title{Retrieve the zone data URLs available to your ICANN CZDAP account}
\usage{
get_zone_data_urls()
}
\value{
character vector of URL paths with embedded api key
}
\description{
Requires \code{ICANN_CZDAP_KEY} to be in the environment (use \code{.Renviron}).
}
\examples{
get_zone_data_urls()
}

