% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/czdaptools.r
\name{get_zones}
\alias{get_zones}
\title{Download all zone files associated with ICANN_CZDAP_KEY.}
\usage{
get_zones(base_zone_dir = "~/Data/zonefiles")
}
\arguments{
\item{base_zone_dir}{base directory where the zone files will be delivered}
}
\description{
Works in parallel (set \code{options(mc.cores=###)}).
}
\details{
This function takes in a base directory location to store ICANN zone files,
creates a subdirectory based on the current date and then tries to
download the zone files in parallel. If not on Windows, use
\code{options(mc.cores=n)} to control the number of parallel jobs.
}

